python/lyrics_audio/requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
syncedlyrics
syrics
python-dotenv